Update axios.js
In this refactoring, we create an instance of `axios` with the base URL set to `http://localhost:3500`, and store it in a constant called `instance`. Finally, we export this constant as the default value of the module.

This refactoring makes the code more readable and less prone to errors. In addition, it uses the default `axios` functions to create the instance, rather than relying on a custom file, making it easy to update your code in the future if the library or instance configuration is changed.